e-modules/engage: simplify zoom function
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@62475 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ E-MODULES-EXTRA/engage/src/e_mod_main.h
@@ -140,7 +140,7 @@ struct _Ng
   unsigned char    changed : 1;
 
   int              pos;  /* mouse position    */
-  int              size; /* current icon size */
+  double           size; /* current icon size */
 
   enum { unzoomed, zooming, zoomed, unzooming } state;
   enum { hidden, hiding, showing, show } hide_state;
